+++ app/src/main/java/com/masonsoft/imsdk/sample/im/FetchSparkMessagePacket.java
@@ -53,7 +53,7 @@ public class FetchSparkMessagePacket extends NotNullTimeoutMessagePacket {
                     }
 
                     if (result.getCode() != 0) {
-                        setErrorCode(result.getCode());
+                        setErrorCode((int) result.getCode());
                         setErrorMessage(result.getMsg());
                         IMLog.e(Objects.defaultObjectTag(this) +
                                 " unexpected. errorCode:%s, errorMessage:%s", result.getCode(), result.getMsg());
