+++ imsdk/src/main/java/com/masonsoft/imsdk/core/IMSessionMessageUploadManager.java
@@ -351,7 +351,7 @@ public class IMSessionMessageUploadManager {
                                 }
 
                                 if (result.getCode() != 0) {
-                                    setErrorCode(result.getCode());
+                                    setErrorCode((int) result.getCode());
                                     setErrorMessage(result.getMsg());
                                     IMLog.e(Objects.defaultObjectTag(ChatSMessagePacket.this) +
                                             " unexpected. errorCode:%s, errorMessage:%s", result.getCode(), result.getMsg());
